extract AssistButton component, fix ui jump
