+++ catalog/app/containers/Bucket/File.js
@@ -6,7 +6,6 @@ import * as React from 'react'
 import { Link, useHistory, useLocation, useParams } from 'react-router-dom'
 import * as M from '@material-ui/core'
 
-import * as Assistant from 'components/Assistant'
 import * as BreadCrumbs from 'components/BreadCrumbs'
 import * as Buttons from 'components/Buttons'
 import * as FileEditor from 'components/FileEditor'
@@ -30,6 +29,7 @@ import parseSearch from 'utils/parseSearch'
 import { up, decode, handleToHttpsUri } from 'utils/s3paths'
 import { readableBytes, readableQuantity } from 'utils/string'
 
+import AssistButton from './AssistButton'
 import FileCodeSamples from './CodeSamples/File'
 import * as AssistantContext from './FileAssistantContext'
 import FileProperties from './FileProperties'
@@ -38,19 +38,6 @@ import Section from './Section'
 import renderPreview from './renderPreview'
 import * as requests from './requests'
 import { useViewModes, viewModeToSelectOption } from './viewModes'
-
-function SummarizeButton() {
-  const assist = Assistant.use()
-  if (!assist) return null
-  const msg = 'Summarize this document'
-  return (
-    <M.IconButton color="primary" onClick={() => assist(msg)} edge="end">
-      <M.Tooltip title="Summarize and chat with AI">
-        <M.Icon>assistant</M.Icon>
-      </M.Tooltip>
-    </M.IconButton>
-  )
-}
 
 const useVersionInfoStyles = M.makeStyles(({ typography }) => ({
   version: {
@@ -514,7 +501,7 @@ export default function File() {
           {BucketPreferences.Result.match(
             {
               // XXX: only show this when the object exists?
-              Ok: ({ ui: { blocks } }) => (blocks.qurator ? <SummarizeButton /> : null),
+              Ok: ({ ui }) => ui.blocks.qurator && <AssistButton edge="end" />,
               _: () => null,
             },
             prefs,
